Move StatChangedEventArgs
